+++ quantify_scheduler/backends/circuit_to_device.py
@@ -14,7 +14,6 @@ from quantify_scheduler.backends.graph_compilation import (
     DeviceCompilationConfig,
     OperationCompilationConfig,
 )
-from quantify_scheduler.backends.qblox.enums import ChannelMode
 from quantify_scheduler.operations.operation import Operation
 from quantify_scheduler.resources import ClockResource
 from quantify_scheduler.schedules.schedule import Schedule, Schedulable, ScheduleBase
@@ -295,9 +294,6 @@ def _set_pulse_and_acquisition_clock(
         operation_info = operation["pulse_info"] + operation["acquisition_info"]
         clocks_used = set([info["clock"] for info in operation_info])
         for clock in clocks_used:
-            # In digital channels clocks are non-existant, so we skip them.
-            if clock == ChannelMode.DIGITAL:
-                continue
             if clock in verified_clocks:
                 continue
             # raises ValueError if no clock found;