+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Tests 
   - Convert transmon-specific hardware configs in test suite and docs to new style. (!972)
 
+- Qblox backend - Allow operations on a digital channel to take a custom clock name as argument. If the clock name is not the default name (`"digital"`), it must be a name that is present in the device configuration, or one that is manually added to the `Schedule` as a `ClockResource`. (!967)
 
 ## 0.20.0 (2024-05-01)